change fetching panther species to fetch from JSON
