+++ src/__test__/utils/pathwayAnalysis/getPantherDBSpecies.test.js
@@ -1,0 +1,8 @@
+import getPantherDBSpecies from 'utils/pathwayAnalysis/getPantherDBSpecies';
+
+describe('getPantherDBSpecies', () => {
+  it('Returns properly formatted', () => {
+    const data = getPantherDBSpecies();
+    expect(data).toMatchSnapshot();
+  });
+});
