--- src/utils/pathwayAnalysis/usePantherDBSpecies.js
@@ -1,32 +1,0 @@
-import { serviceUrls } from 'utils/pathwayAnalysis/pathwayConstants';
-import useSWR from 'swr';
-
-const formatOuput = (data) => {
-  const { genome } = data.search.output.genomes;
-
-  const speciesList = genome.map((species) => ({
-    label: species.long_name,
-    value: species.short_name,
-  }));
-
-  return speciesList;
-};
-
-const usePantherDBSpecies = () => {
-  const { data, error } = useSWR(
-    serviceUrls.pantherDBSpecies,
-    async (url) => {
-      const response = await fetch(url);
-      const responseData = await response.json();
-      return responseData;
-    },
-  );
-
-  const formattedData = data ? formatOuput(data) : [];
-  return {
-    data: formattedData,
-    error,
-  };
-};
-
-export default usePantherDBSpecies;
